% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtrunc.R
\name{dtrunc}
\alias{dtrunc}
\title{Truncated density generation for any univariate distribution in the namespace}
\usage{
dtrunc(a = -Inf, b = Inf, dist = "norm", x, args = list(mean = 0, sd = 1))
}
\arguments{
\item{a}{left-truncation point.}

\item{b}{right-truncation point}

\item{dist}{name of the univariate distribution (e.g., `norm`, `lnorm`, `binom`,
`gompertz`, `gamma`).}

\item{x}{vector of quantiles.}

\item{args}{a named list of parameter vectors corresponding to the distribution
(e.g., `list(mean = 10, sd = 100)` for `norm`)}
}
\value{
Generates density
}
\description{
`dtrunc` generates density for any truncated univariate distribution that is present in the namespace
}
\examples{
dtrunc(a = 0, dist = "norm", x = 1, args = list(mean = 10, sd = 10))
dtrunc(a = 2, dist = "binom", x = 3, args = list(size = 10, prob = 0.2))
}
