% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrunc.R
\name{ptrunc}
\alias{ptrunc}
\title{Gives the truncated distribution function for any univariate distribution in the namespace}
\usage{
ptrunc(
  a = -Inf,
  b = Inf,
  dist = "norm",
  q,
  args = list(mean = 0, sd = 1),
  lower.tail = TRUE
)
}
\arguments{
\item{a}{left-truncation point.}

\item{b}{right-truncation point}

\item{dist}{name of the univariate distribution (e.g., `norm`, `lnorm`, `binom`,
`gompertz`, `gamma`).}

\item{q}{vector of quantiles.}

\item{args}{a named list of parameter vectors corresponding to the distribution
(e.g., `list(mean = 10, sd = 100)` for `norm`)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}
}
\value{
generates the distribution function
}
\description{
`ptrunc` generates distribution function for any truncated univariate distribution that is present in the namespace
}
\examples{
ptrunc(a = 0, dist = "norm", q = 1.96, args = list(mean = 10, sd = 10))
ptrunc(a = 2, dist = "binom", q = 6, args = list(size = 10, prob = 0.2))
}
