% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtrunc.R
\name{qtrunc}
\alias{qtrunc}
\title{Gives the truncated quantile function for any univariate distribution in the namespace}
\usage{
qtrunc(
  a = -Inf,
  b = Inf,
  dist = "norm",
  p,
  args = list(mean = 0, sd = 1),
  lower.tail = TRUE
)
}
\arguments{
\item{a}{left-truncation point.}

\item{b}{right-truncation point}

\item{dist}{name of the univariate distribution (e.g., `norm`, `lnorm`, `binom`,
`gompertz`, `gamma`).}

\item{p}{vector of probabilities.}

\item{args}{a named list of parameter vectors corresponding to the distribution
(e.g., `list(mean = 10, sd = 100)` for `norm`)}

\item{lower.tail}{logical; if TRUE (default), probabilities are \eqn{P[X \leq x]} otherwise, \eqn{P[X > x]}.}
}
\value{
gives the distribution function
}
\description{
`qtrunc` generates quantile function for any truncated univariate distribution that is present in the namespace
}
\examples{
qtrunc(a = 0, dist = "norm", p = 0.95, args = list(mean = 10, sd = 10))
qtrunc(a = 2, dist = "binom", p = 0.5, args = list(size = 10, prob = 0.2))
}
