% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\title{Random truncated generation for any univariate distribution in the namespace}
\usage{
rtrunc(
  n = 1,
  a = -Inf,
  b = Inf,
  dist = "gompertz",
  args = list(shape = 1, rate = 1)
)
}
\arguments{
\item{n}{number of observations. If length(n) > 1, the length is taken to be the number required.}

\item{a}{left-truncation point.}

\item{b}{right-truncation point}

\item{dist}{name of the univariate distribution (e.g., `norm`, `lnorm`, `binom`,
`gompertz`, `gamma`).}

\item{args}{a named list of parameter vectors corresponding to the distribution
(e.g., `list(mean = 10, sd = 100)` for `norm`)}
}
\value{
Generates random deviates. The length of the result is determined by n.
The numerical arguments other than n are recycled to the length of the result.
}
\description{
`rtrunc` generates random numbers for any truncated univariate distribution that is present in the namespace
}
\examples{
rtrunc(n = 10, a = 0, dist = "norm", args = list(mean = 10, sd = 10))
rtrunc(n = 10, a = 2, dist = "binom", args = list(size = 10, prob = 0.2))
}
