% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc-package.R
\docType{package}
\name{trunc-package}
\alias{trunc}
\alias{trunc-package}
\title{trunc: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: swarnendu chatterjee \email{swarnendu.stat@gmail.com}

}
\keyword{internal}
