% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trunc-package.R
\docType{package}
\name{trunc-package}
\alias{trunc}
\alias{trunc-package}
\title{trunc: Truncated Distribution (Title Case)}
\description{
Density, distribution function, quantile function and random generation for truncated version of any distribution in the namespace.
}
\author{
\strong{Maintainer}: swarnendu chatterjee \email{swarnendu.stat@gmail.com}

}
\keyword{internal}
